Run isort on shoebot/data/__init__, basecolor, img, typography Move RGB, colour constants to basecolor.
